simplify

+++ loader.js
@@ -35,7 +35,7 @@ const dll = ffi.Library(
     'acpi_read32': ['Uint32', ['Uint32']],
     'acpi_write8': ['void', ['Uint32', 'Uint8']],
     'acpi_write16': ['void', ['Uint32', 'Uint16']],
-    'acpi_get_timer': ['Uint64', ['Uint64']],
+    'acpi_get_timer': ['Uint32', ['Uint64']],
     'acpi_timer': ['bool', ['Uint64']],
     'acpi_get_result': ['int', []],
   }
